Added  a simple dashboard and admin manager page
In this commit, I've started adding the backend part of the project. First came about the process of making the dashboard, a simple menu to access the different pages, and then the admin manager page where I've added the functionality to add, update and delete an admin. Plus the CSS to make it look good :)